Fix navigation but on NewCard screen
Previously this was causing the new card screen to navigate instead of
going back, which was causing the "Back Arrow" naivgation to take you
back to the NewCard screen after you created a new card. We want that
back arrow on the DeckDetails screen to always take you back to the list
of Decks.